refactor: move LdtkProject to its own module and re-export it
