+++ src/assets/mod.rs
@@ -1,6 +1,6 @@
 //! Assets and AssetLoaders for loading ldtk files.
 
-use crate::{ldtk, resources::LevelSelection};
+use crate::ldtk;
 use bevy::{
     asset::{AssetLoader, AssetPath, LoadContext, LoadedAsset},
     prelude::*,
@@ -8,30 +8,13 @@ use bevy::{
     render::render_resource::{Extent3d, TextureDimension, TextureFormat},
     utils::BoxedFuture,
 };
-use std::{collections::HashMap, path::Path};
+use std::path::Path;
 
-#[allow(unused_imports)]
-use crate::components::LdtkWorldBundle;
+mod ldtk_project;
+pub use ldtk_project::{LdtkLoader, LdtkProject};
 
 fn ldtk_path_to_asset_path<'b>(ldtk_path: &Path, rel_path: &str) -> AssetPath<'b> {
     ldtk_path.parent().unwrap().join(Path::new(rel_path)).into()
-}
-
-/// Main asset for loading ldtk files.
-///
-/// Load your ldtk project with the asset server, then insert the handle into the
-/// [LdtkWorldBundle].
-#[derive(Clone, TypeUuid)]
-#[uuid = "ecfb87b7-9cd9-4970-8482-f2f68b770d31"]
-pub struct LdtkProject {
-    /// Raw ldtk project data.
-    pub project: ldtk::LdtkJson,
-    /// Map from tileset uids to image handles for the loaded tileset.
-    pub tileset_map: HashMap<i32, Handle<Image>>,
-    /// Map from level iids to level handles.
-    pub level_map: HashMap<String, Handle<LdtkLevel>>,
-    /// Image used for rendering int grid colors.
-    pub int_grid_image_handle: Option<Handle<Image>>,
 }
 
 impl ldtk::Definitions {
@@ -68,129 +51,6 @@ impl ldtk::LdtkJson {
         self.levels
             .iter()
             .chain(self.worlds.iter().flat_map(|w| &w.levels))
-    }
-}
-
-impl LdtkProject {
-    pub fn world_height(&self) -> i32 {
-        let mut world_height = 0;
-        for level in self.iter_levels() {
-            world_height = world_height.max(level.world_y + level.px_hei);
-        }
-
-        world_height
-    }
-
-    /// Get an iterator of all the levels in the LDtk file.
-    ///
-    /// This abstraction avoids compatibility issues between pre-multi-world and post-multi-world
-    /// LDtk projects.
-    ///
-    /// Note: the returned levels are the ones existent in the [LdtkProject].
-    /// These levels will have "incomplete" data if you use LDtk's external levels feature.
-    /// To always get full level data, you'll need to access `Assets<LdtkLevel>`.
-    pub fn iter_levels(&self) -> impl Iterator<Item = &ldtk::Level> {
-        self.project.iter_levels()
-    }
-
-    /// Find a particular level using a [LevelSelection].
-    ///
-    /// Note: the returned level is the one existent in the [LdtkProject].
-    /// This level will have "incomplete" data if you use LDtk's external levels feature.
-    /// To always get full level data, you'll need to access `Assets<LdtkLevel>`.
-    pub fn get_level(&self, level_selection: &LevelSelection) -> Option<&ldtk::Level> {
-        self.iter_levels()
-            .enumerate()
-            .find(|(i, l)| level_selection.is_match(i, l))
-            .map(|(_, l)| l)
-    }
-}
-
-#[derive(Copy, Clone, Debug, Default)]
-pub struct LdtkLoader;
-
-impl AssetLoader for LdtkLoader {
-    fn load<'a>(
-        &'a self,
-        bytes: &'a [u8],
-        load_context: &'a mut LoadContext,
-    ) -> BoxedFuture<'a, anyhow::Result<()>> {
-        Box::pin(async move {
-            let project: ldtk::LdtkJson = serde_json::from_slice(bytes)?;
-
-            let mut external_level_paths = Vec::new();
-            let mut level_map = HashMap::new();
-            let mut background_images = Vec::new();
-            if project.external_levels {
-                for level in project.iter_levels() {
-                    if let Some(external_rel_path) = &level.external_rel_path {
-                        let asset_path =
-                            ldtk_path_to_asset_path(load_context.path(), external_rel_path);
-
-                        external_level_paths.push(asset_path.clone());
-                        level_map.insert(level.iid.clone(), load_context.get_handle(asset_path));
-                    }
-                }
-            } else {
-                for level in project.iter_levels() {
-                    let label = level.identifier.as_ref();
-
-                    let mut background_image = None;
-                    if let Some(rel_path) = &level.bg_rel_path {
-                        let asset_path = ldtk_path_to_asset_path(load_context.path(), rel_path);
-                        background_images.push(asset_path.clone());
-                        background_image = Some(load_context.get_handle(asset_path));
-                    }
-
-                    let ldtk_level = LdtkLevel {
-                        level: level.clone(),
-                        background_image,
-                    };
-                    let level_handle =
-                        load_context.set_labeled_asset(label, LoadedAsset::new(ldtk_level));
-
-                    level_map.insert(level.iid.clone(), level_handle);
-                }
-            }
-
-            let mut tileset_rel_paths = Vec::new();
-            let mut tileset_map = HashMap::new();
-            for tileset in &project.defs.tilesets {
-                if let Some(tileset_path) = &tileset.rel_path {
-                    let asset_path = ldtk_path_to_asset_path(load_context.path(), tileset_path);
-
-                    tileset_rel_paths.push(asset_path.clone());
-                    tileset_map.insert(tileset.uid, load_context.get_handle(asset_path));
-                } else if tileset.embed_atlas.is_some() {
-                    warn!("Ignoring LDtk's Internal_Icons. They cannot be displayed due to their license.");
-                } else {
-                    let identifier = &tileset.identifier;
-                    warn!("{identifier} tileset cannot be loaded, it has a null relative path.");
-                }
-            }
-
-            let int_grid_image_handle = project.defs.create_int_grid_image().map(|image| {
-                load_context.set_labeled_asset("int_grid_image", LoadedAsset::new(image))
-            });
-
-            let ldtk_asset = LdtkProject {
-                project,
-                tileset_map,
-                level_map,
-                int_grid_image_handle,
-            };
-            load_context.set_default_asset(
-                LoadedAsset::new(ldtk_asset)
-                    .with_dependencies(tileset_rel_paths)
-                    .with_dependencies(external_level_paths)
-                    .with_dependencies(background_images),
-            );
-            Ok(())
-        })
-    }
-
-    fn extensions(&self) -> &[&str] {
-        &["ldtk"]
     }
 }
